Rename keyboard component
