+++ app/components/Views/confirmations/components/deposit-keyboard/index.ts
@@ -1,0 +1,1 @@
+export * from './deposit-keyboard';
