--- app/components/Views/confirmations/components/edit-amount-keyboard/index.ts
@@ -1,1 +1,0 @@
-export * from './edit-amount-keyboard';
